requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
confluent-kafka==2.3.0
redis==5.0.4
django-redis==5.4.0
django-ipware==6.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1